Deactivated most of IterSolvers, but now the Parent class should work fine, with redesigned test.

+++ Tests/IterSolvers/unitt_64.cpp
@@ -15,7 +15,7 @@ int main() {
     double tol = 1e-6;
 
     // Create an instance of TestSolver
-    TestSolver<uint32_t,float> tSolv(clientComm, arrSize_l, maxIters, tol);
+    TestSolver<uint32_t,double> tSolv(clientComm, arrSize_l, maxIters, tol);
 
     // Setup test
     PUSH_RANGE("Setup Test",5);
@@ -26,18 +26,18 @@ int main() {
         dim3 kBlock(TILE_SIZE,1,1);
         DeviceUtils::Stream_t kStream;
         DeviceUtils::StreamCreate(&kStream);
-        float* d_inicond_l = DeviceMemory<uint32_t,float>::deviceCalloc(arrSize_l);
-        float* d_rhs_l = DeviceMemory<uint32_t,float>::deviceCalloc(arrSize_l);
-        DeviceUtils::launchKernel(set_array<uint32_t, float>, kGrid, kBlock, kStream, d_inicond_l, 1.0f, arrSize_l);
-        DeviceUtils::launchKernel(set_array<uint32_t, float>, kGrid, kBlock, kStream, d_rhs_l, 3.0f, arrSize_l);
+        double* d_inicond_l = DeviceMemory<uint32_t,double>::deviceCalloc(arrSize_l);
+        double* d_rhs_l = DeviceMemory<uint32_t,double>::deviceCalloc(arrSize_l);
+        DeviceUtils::launchKernel(set_array<uint32_t, double>, kGrid, kBlock, kStream, d_inicond_l, 1.0, arrSize_l);
+        DeviceUtils::launchKernel(set_array<uint32_t, double>, kGrid, kBlock, kStream, d_rhs_l, 3.0, arrSize_l);
         DeviceUtils::StreamSynchronize(kStream);
         tSolv.setup(d_inicond_l, d_rhs_l);
     #else
-        float* inicond_l = (float*)calloc(arrSize_l, sizeof(float));
-        float* rhs_l = (float*)calloc(arrSize_l, sizeof(float));
+        double* inicond_l = (double*)calloc(arrSize_l, sizeof(double));
+        double* rhs_l = (double*)calloc(arrSize_l, sizeof(double));
         for (uint32_t i = 0; i < arrSize_l; ++i) {
-            inicond_l[i] = 1.0f;
-            rhs_l[i] = 3.0f;
+            inicond_l[i] = 1.0;
+            rhs_l[i] = 3.0;
         }
         tSolv.setup(inicond_l, rhs_l);
     #endif
@@ -49,11 +49,11 @@ int main() {
 
     // Retrieve the solution
     PUSH_RANGE("Retrieve Solution",5);
-    float* h_res = (float*)calloc(arrSize_l, sizeof(float));
+    double* h_res = (double*)calloc(arrSize_l, sizeof(double));
     #if defined(USE_GPU)
-        float* d_res = DeviceMemory<uint32_t,float>::deviceCalloc(arrSize_l);
+        double* d_res = DeviceMemory<uint32_t,double>::deviceCalloc(arrSize_l);
         tSolv.getSolution(d_res);
-        DeviceMemory<uint32_t,float>::copyDeviceToHost(arrSize_l, d_res, h_res);
+        DeviceMemory<uint32_t,double>::copyDeviceToHost(arrSize_l, d_res, h_res);
     #else
         tSolv.getSolution(h_res);
     #endif
@@ -62,7 +62,7 @@ int main() {
     // Check result
     PUSH_RANGE("Check Result",5);
     for (uint32_t i = 0; i < arrSize_l; ++i) {
-        if (h_res[i] != 2.0f) {
+        if (h_res[i] != 2.0) {
             std::cerr << "Test failed at index " << i << ": expected 2.0, got " << h_res[i] << std::endl;
             exit(EXIT_FAILURE);
         }
